refactor init_optimizers
