+++ pytorch_lightning/plugins/training_type/training_type_plugin.py
@@ -105,9 +105,7 @@ class TrainingTypePlugin(ABC):
         """
         if trainer.state.fn not in (TrainerFn.FITTING, TrainerFn.TUNING):
             return
-        optimizers, lr_schedulers, optimizer_frequencies = self.init_optimizers(
-            trainer=trainer, model=self.lightning_module
-        )
+        optimizers, lr_schedulers, optimizer_frequencies = self.lightning_module.init_optimizers_and_lr_schedulers()
         self.optimizers = optimizers
         self.lr_schedulers = lr_schedulers
         self.optimizer_frequencies = optimizer_frequencies
@@ -376,9 +374,6 @@ class TrainingTypePlugin(ABC):
             dataloader: iterable. Ideally of type: :class:`torch.utils.data.DataLoader`
         """
         return dataloader
-
-    def init_optimizers(self, trainer: "pl.Trainer", model: "pl.LightningModule"):
-        return trainer.init_optimizers(model)
 
     @property
     def restore_checkpoint_after_pre_dispatch(self) -> bool:
